requirements.txt
streamlit
opencv-python
numpy
pillow
tensorflow
scikit-learn